requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Pillow==10.1.0
